yet another proposed JS reload to the login page
Fixed errors on GCE, probably a good idea to add to the entire project as it will work with any system/config. I found adding /login helped the issues with redirecting as it checks if the user is logged in and will re-direct them accordingly. Cuts down on redirects which is not only bad, but can confuse some browsers when switching from http->https